More reordering tests
