+++ .gitignore
@@ -3,3 +3,4 @@ dist/
 *.pyc
 django_ratings.egg-info/
 *.egg
+.tox/
