+++ djangoratings/default_settings.py
