+++ djangoratings/models.py
@@ -4,11 +4,11 @@ from django.db import models
 from django.contrib.contenttypes.models import ContentType
 from django.contrib.auth.models import User
 import django
-if (hasattr(django,"version") and django.version > 1.8) or (hasattr(django,"get_version") and django.get_version()):
+if (hasattr(django, "version") and django.version > 1.8) or (hasattr(django, "get_version") and django.get_version()):
     from django.contrib.contenttypes.fields import GenericForeignKey
 else:
     from django.contrib.contenttypes.generic import GenericForeignKey
-    
+
 
 try:
     from django.utils.timezone import now
@@ -24,14 +24,13 @@ class Vote(models.Model):
     key = models.CharField(max_length=32)
     score = models.IntegerField()
     user = models.ForeignKey(User, blank=True, null=True, related_name="votes")
-    ip_address = models.GenericIPAddressField()
-    ip_address      = models.GenericIPAddressField() if hasattr(models,"GenericIPAddressField") else models.IPAddressField() 
+    ip_address = models.GenericIPAddressField() if hasattr(models, "GenericIPAddressField") else models.IPAddressField() 
     cookie = models.CharField(max_length=32, blank=True, null=True)
     date_added = models.DateTimeField(default=now, editable=False)
     date_changed = models.DateTimeField(default=now, editable=False)
     objects = VoteManager()
 
-    content_object  = GenericForeignKey()
+    content_object = GenericForeignKey()
 
     class Meta:
         unique_together = (('content_type', 'object_id', 'key', 'user', 'ip_address', 'cookie'))
@@ -55,13 +54,13 @@ class Vote(models.Model):
         return '.'.join(ip)
     partial_ip_address = property(partial_ip_address)
 
+
 class Score(models.Model):
-    content_type    = models.ForeignKey(ContentType)
-    object_id       = models.PositiveIntegerField()
-    key             = models.CharField(max_length=32)
-    score           = models.IntegerField()
-    votes           = models.PositiveIntegerField()
-    
+    content_type = models.ForeignKey(ContentType)
+    object_id = models.PositiveIntegerField()
+    key = models.CharField(max_length=32)
+    score = models.IntegerField()
+    votes = models.PositiveIntegerField()
     content_object  = GenericForeignKey()
 
     class Meta:
@@ -70,30 +69,30 @@ class Score(models.Model):
     def __unicode__(self):
         return u"%s scored %s with %s votes" % (self.content_object, self.score, self.votes)
 
+
 class SimilarUser(models.Model):
-    from_user       = models.ForeignKey(User, related_name="similar_users")
-    to_user         = models.ForeignKey(User, related_name="similar_users_from")
-    agrees          = models.PositiveIntegerField(default=0)
-    disagrees       = models.PositiveIntegerField(default=0)
-    exclude         = models.BooleanField(default=False)
-    
-    objects         = SimilarUserManager()
-    
+    from_user = models.ForeignKey(User, related_name="similar_users")
+    to_user = models.ForeignKey(User, related_name="similar_users_from")
+    agrees = models.PositiveIntegerField(default=0)
+    disagrees = models.PositiveIntegerField(default=0)
+    exclude = models.BooleanField(default=False)
+    objects = SimilarUserManager()
+
     class Meta:
         unique_together = (('from_user', 'to_user'),)
 
     def __unicode__(self):
         print u"%s %s similar to %s" % (self.from_user, self.exclude and 'is not' or 'is', self.to_user)
 
+
 class IgnoredObject(models.Model):
-    user            = models.ForeignKey(User)
-    content_type    = models.ForeignKey(ContentType)
-    object_id       = models.PositiveIntegerField()
-    
-    content_object  = GenericForeignKey()
-    
+    user = models.ForeignKey(User)
+    content_type = models.ForeignKey(ContentType)
+    object_id = models.PositiveIntegerField()
+    content_object = GenericForeignKey()
+
     class Meta:
         unique_together = (('content_type', 'object_id'),)
-    
+
     def __unicode__(self):
         return self.content_object
