+++ requirements-test.txt
@@ -1,2 +1,3 @@
 dj-database-url==0.3.0
 psycopg2==2.7.1
+django==1.8.18
