+++ tox.ini
@@ -1,13 +1,11 @@
 [tox]
-envlist = {py27}-django{17,18,19,110,111}
+envlist = {py27}-django{19,110,111}
 
 [testenv]
 passenv = *
 basepython =
     py27: python2.7
 deps =
-    django17: Django>=1.7,<1.8
-    django18: Django>=1.8,<1.9
     django19: Django>=1.9,<1.10
     django110: Django>=1.10,<1.11
     django111: Django>=1.11,<1.12
